changed game rules; fixed some bugs; added debug
changed game rules. Now, the good cells are generated not distinctly, but are connected in their rows and the game modes are defined not by chance of some cell being good, but by the constant number of good cells in each row

fixed some bugs:
- you could start a new game while having another active one
- when calculating income, some debug lines were output to cout
- when the game is over (win or lost), the balance was not yet updated when being output

added debug: if variable DEBUG from the 5th line is set to true, you can use the "secret" 42 option when playing the game to output the whole board (with all the good cells highlighted). Otherwise, the game, hopefully, acts normally